Clean up and merge D3 sensors and their names
