+++ custom_components/ef_ble/sensor.py
@@ -182,7 +182,7 @@ SENSOR_TYPES: dict[str, SensorEntityDescription] = {
     ),
     "usbc_output_power": SensorEntityDescription(
         key="usbc_output_power",
-        name="USC C Output Power",
+        name="USB C Output Power",
         icon="mdi:usb-c-port",
         native_unit_of_measurement=UnitOfPower.WATT,
         device_class=SensorDeviceClass.POWER,
@@ -200,7 +200,7 @@ SENSOR_TYPES: dict[str, SensorEntityDescription] = {
     ),
     "usba_output_power": SensorEntityDescription(
         key="usba_output_power",
-        name="USC A Output Power",
+        name="USB A Output Power",
         icon="mdi:usb-port",
         native_unit_of_measurement=UnitOfPower.WATT,
         device_class=SensorDeviceClass.POWER,
@@ -215,6 +215,22 @@ SENSOR_TYPES: dict[str, SensorEntityDescription] = {
         state_class=SensorStateClass.TOTAL_INCREASING,
         suggested_display_precision=3,
         suggested_unit_of_measurement=UnitOfEnergy.KILO_WATT_HOUR,
+    ),
+    "usbc2_output_power": SensorEntityDescription(
+        key="usbc2_output_power",
+        name="USB C (2) Output Power",
+        icon="mdi:usb-c-port",
+        native_unit_of_measurement=UnitOfPower.WATT,
+        device_class=SensorDeviceClass.POWER,
+        state_class=SensorStateClass.MEASUREMENT,
+    ),
+    "usba2_output_power": SensorEntityDescription(
+        key="usba2_output_power",
+        name="USB A (2) Output Power",
+        icon="mdi:usb-port",
+        native_unit_of_measurement=UnitOfPower.WATT,
+        device_class=SensorDeviceClass.POWER,
+        state_class=SensorStateClass.MEASUREMENT,
     ),
     "battery_input_power": SensorEntityDescription(
         key="battery_input_power",
@@ -240,9 +256,9 @@ SENSOR_TYPES: dict[str, SensorEntityDescription] = {
         native_unit_of_measurement=UnitOfTemperature.CELSIUS,
         entity_registry_enabled_default=False,
     ),
-    "dc_port_1_input_power": SensorEntityDescription(
-        key="dc_port_input_power_left",
-        name="DC Port Input Power Left",
+    "dc_port_input_power": SensorEntityDescription(
+        key="dc_port_input_power",
+        name="DC Port Input Power",
         icon="mdi:current-dc",
         device_class=SensorDeviceClass.POWER,
         state_class=SensorStateClass.MEASUREMENT,
@@ -250,8 +266,8 @@ SENSOR_TYPES: dict[str, SensorEntityDescription] = {
         suggested_display_precision=2,
     ),
     "dc_port_2_input_power": SensorEntityDescription(
-        key="dc_port_input_power_right",
-        name="DC Port Input Power Right",
+        key="dc_port_input_power_2",
+        name="DC Port (2) Input Power",
         icon="mdi:current-dc",
         device_class=SensorDeviceClass.POWER,
         state_class=SensorStateClass.MEASUREMENT,
@@ -259,13 +275,19 @@ SENSOR_TYPES: dict[str, SensorEntityDescription] = {
         suggested_display_precision=2,
     ),
     "dc_port_state": SensorEntityDescription(
-        key="dc_port_state",
+        key="dc_port_1_state",
         name="DC Input Port State",
         icon="mdi:current-dc",
         device_class=SensorDeviceClass.ENUM,
     ),
+    "dc_port_2_state": SensorEntityDescription(
+        key="dc_port_2_state",
+        name="DC Input Port (2) State ",
+        icon="mdi:current-dc",
+        device_class=SensorDeviceClass.ENUM,
+    ),
     "solar_input_power": SensorEntityDescription(
-        key="solar_input_power",
+        key="input_power_solar",
         name="Solar Power",
         icon="mdi:solar-power-variant-outline",
         native_unit_of_measurement=UnitOfPower.WATT,
@@ -273,69 +295,14 @@ SENSOR_TYPES: dict[str, SensorEntityDescription] = {
         state_class=SensorStateClass.MEASUREMENT,
         entity_registry_enabled_default=False,
     ),
-    "dc_port_1_state": SensorEntityDescription(
-        key="dc_port_state_left",
-        name="DC Input Port State Left",
-        icon="mdi:current-dc",
-        device_class=SensorDeviceClass.ENUM,
-        state_class=None,
-    ),
-    "dc_port_2_state": SensorEntityDescription(
-        key="dc_port_state_right",
-        name="DC Input Port State Right",
-        icon="mdi:current-dc",
-        device_class=SensorDeviceClass.ENUM,
-        state_class=None,
-    ),
-    "solar_input_power_1": SensorEntityDescription(
-        key="input_power_solar_left",
-        name="Solar Power Left",
-        icon="mdi:solar-power-variant-outline",
-        native_unit_of_measurement=UnitOfPower.WATT,
-        device_class=SensorDeviceClass.POWER,
-        state_class=SensorStateClass.MEASUREMENT,
-        entity_registry_enabled_default=False,
-    ),
     "solar_input_power_2": SensorEntityDescription(
-        key="input_power_solar_right",
-        name="Solar Power Right",
+        key="input_power_solar_2",
+        name="Solar Power (2)",
         icon="mdi:solar-power-variant-outline",
         native_unit_of_measurement=UnitOfPower.WATT,
         device_class=SensorDeviceClass.POWER,
         state_class=SensorStateClass.MEASUREMENT,
         entity_registry_enabled_default=False,
-    ),
-    "usbc1_output_power": SensorEntityDescription(
-        key="usbc1_output_power",
-        name="USB C Left Output Power",
-        icon="mdi:usb-c-port",
-        native_unit_of_measurement=UnitOfPower.WATT,
-        device_class=SensorDeviceClass.POWER,
-        state_class=SensorStateClass.MEASUREMENT,
-    ),
-    "usbc2_output_power": SensorEntityDescription(
-        key="usbc2_output_power",
-        name="USB C Right Output Power",
-        icon="mdi:usb-c-port",
-        native_unit_of_measurement=UnitOfPower.WATT,
-        device_class=SensorDeviceClass.POWER,
-        state_class=SensorStateClass.MEASUREMENT,
-    ),
-    "usba1_output_power": SensorEntityDescription(
-        key="usba_output_power",
-        name="USB A Left Output Power",
-        icon="mdi:usb-port",
-        native_unit_of_measurement=UnitOfPower.WATT,
-        device_class=SensorDeviceClass.POWER,
-        state_class=SensorStateClass.MEASUREMENT,
-    ),
-    "usba2_output_power": SensorEntityDescription(
-        key="usba_output_power",
-        name="USB A Right Output Power",
-        icon="mdi:usb-port",
-        native_unit_of_measurement=UnitOfPower.WATT,
-        device_class=SensorDeviceClass.POWER,
-        state_class=SensorStateClass.MEASUREMENT,
     ),
 }
